Pack flag for /v2/observations
Require the 'pack' flag for Protobuf output. Accepted values: 't' and 'f', but currently only 'f' is supported.

This flag ensures we can add the packed protobuf output format without breaking compatibility in the future.
If we decide to use packed or unpacked as the default, we can make this flag optional in the future.